Support translucent nodes and edges
