+++ src/utils/utils.js
@@ -1,5 +1,3 @@
-import { TextureLoader } from 'three'
-
 function prettyPrint(number) {
 
     if (typeof number !== "number") {
@@ -9,43 +7,6 @@ function prettyPrint(number) {
 
     const integerPart = Math.trunc(number)
     return integerPart.toLocaleString()
-}
-
-/**
- * Loads a texture asynchronously
- * @param {string} url - The URL of the texture to load
- * @param {Object} options - Optional texture loading options
- * @param {boolean} options.flipY - Whether to flip the texture vertically (default: true)
- * @param {boolean} options.generateMipmaps - Whether to generate mipmaps (default: true)
- * @returns {Promise<THREE.Texture>} A promise that resolves with the loaded texture
- */
-async function loadTexture(url, options = {}) {
-    const loader = new TextureLoader()
-    
-    return new Promise((resolve, reject) => {
-        loader.load(
-            url,
-            (texture) => {
-                // Apply default options
-                texture.flipY = options.flipY !== undefined ? options.flipY : true
-                texture.generateMipmaps = options.generateMipmaps !== undefined ? options.generateMipmaps : true
-                
-                // Apply any additional options
-                Object.entries(options).forEach(([key, value]) => {
-                    if (key !== 'flipY' && key !== 'generateMipmaps') {
-                        texture[key] = value
-                    }
-                })
-                
-                resolve(texture)
-            },
-            undefined, // onProgress callback not implemented
-            (error) => {
-                console.error(`Error loading texture from ${url}:`, error)
-                reject(error)
-            }
-        )
-    })
 }
 
 /**
@@ -88,4 +49,4 @@ function ingestData(json, genomicService, geometryManager) {
     geometryManager.createGeometry(json)
 }
 
-export { prettyPrint, loadTexture, loadPath, ingestData }
+export { prettyPrint, loadPath, ingestData }
